Improve populates and controllers functionallyties
 - improve populate on user model
 - improve populate on immersion model
 - improve create and delete on equipment controller
 - improve create and delete on immersion controller

+++ backend/src/controllers/equipmentController.js
@@ -3,10 +3,31 @@ const Equipment = require('../models/equipmentModel');
 const User = require('../models/userModel');
 
 async function updateUserEquipment(equipmentID, { userId }) {
-  debug('updateUserEquipment');
-  debug(equipmentID);
-  debug(userId);
-  const currentUser = await User.findById(userId);
+  try {
+    debug('updateUserEquipment');
+    const currentUser = await User.findById(userId);
+    if (Object.keys(currentUser).length > 0) {
+      await User.findByIdAndUpdate(userId, { $addToSet: { equipment: equipmentID } });
+      return true;
+    }
+    return false;
+  } catch (error) {
+    return false;
+  }
+}
+
+async function deleteUserEquipment(equipmentID, { userId }) {
+  try {
+    debug('deleteUserEquipment');
+    const currentUser = await User.findById(userId);
+    if (Object.keys(currentUser).length > 0) {
+      await User.findByIdAndUpdate(userId, { $pull: { equipment: { $in: [equipmentID] } } });
+      return true;
+    }
+    return false;
+  } catch (error) {
+    return false;
+  }
 }
 
 async function getEquipment(req, res) {
@@ -38,8 +59,12 @@ async function setEquipment(req, res) {
     }
     const newEquipment = await Equipment.create(req.body);
     // eslint-disable-next-line no-underscore-dangle
-    updateUserEquipment(newEquipment._id, req.query);
-    res.status(201);
+    const updateUser = updateUserEquipment(newEquipment._id, req.query);
+    if (updateUser) {
+      res.status(201);
+    } else {
+      res.status(404);
+    }
     return res.json(newEquipment);
   } catch (error) {
     res.status(500);
@@ -80,8 +105,18 @@ async function updateEquipmentById(req, res) {
 async function deleteEquipmentById(req, res) {
   try {
     debug('deleteEquipmentById');
+    if (req.body.userId === undefined) {
+      res.status(400);
+      return res.send({ message: 'No user on body' });
+    }
     const deletedEquipment = await Equipment.findByIdAndRemove(req.params.equipmentId);
-    res.status(200);
+    // eslint-disable-next-line no-underscore-dangle
+    const updateUser = deleteUserEquipment(req.params.equipmentId, req.body);
+    if (updateUser) {
+      res.status(200);
+    } else {
+      res.status(404);
+    }
     return res.json(deletedEquipment);
   } catch (error) {
     res.status(500);
